requirements.txt
flask==3.0.3
PyMuPDF==1.24.11
zxing-cpp==2.3.0
pillow==10.3.0